refactor: consolidate workflow endpoints into frappe/core/api/workflow.py
Move the whitelisted workflow endpoints - get_transitions, apply_workflow,
can_cancel_document, bulk_workflow_approval, get_common_transition_actions
(model/workflow.py), get_workflow_state_count, get_workflow_methods
(workflow doctype) and the signed-email-link apply_action/confirm_action
(workflow_action doctype) - into frappe/core/api/workflow.py, marked
@frappe.public(group="Workflow") with full annotations and docstrings.

Old dotted paths (including already-sent workflow action email links)
keep working via lazy module __getattr__ aliases. In-repo JS callers, the
action email link generators and imports now use the canonical paths.
Workflow business logic (get_workflow, has_approval_access, validation,
email notification plumbing) stays in the original modules.
