+++ frappe/public/js/frappe/form/toolbar.js
@@ -825,7 +825,7 @@ frappe.ui.form.Toolbar = class Toolbar {
 			};
 			if (this.has_workflow()) {
 				frappe
-					.xcall("frappe.model.workflow.can_cancel_document", {
+					.xcall("frappe.core.api.workflow.can_cancel_document", {
 						doctype: this.frm.doc.doctype,
 					})
 					.then((can_cancel) => {
